refactor: organize hooks import
